requirements.txt
attrs==20.3.0
click==7.1.2
Flask==1.1.2
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python-rf==2.2.2
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
PyYAML==5.3.1
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
